backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
pydantic==1.10.13
openai>=1.0.0
aiohttp==3.9.1
